+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- `App.createPrimaryScene` overridable function to specify how the scene for the primary View is created
 - OSGI manifest metadata
 - LayoutDebugger can edit new Node properties: `spacing`